requirements.txt
attrs>=19.1
certvalidator
cryptoparser>=0.7.1
enum34==1.1.6;python_version<"3.4"
py2-ipaddress;python_version<"3.4"
python-dateutil
six
typing;python_version<"3.5"
urllib3
